+++ library/src/ts/GifWriter.test.ts
@@ -1,22 +1,20 @@
-import {GifWriter, IndexedColorImage} from "../main/GifWriter";
-import t from "./test_common";
+import {GifWriter, IndexedColorImage, IOutputStream} from "./GifWriter";
+import t from "./__test__/test_common";
 
-(function () {
-"use strict";
-
-function createOutputStream() {
-    return {
-        buffer: <string[]>[],
-        writeByte: function (b: number) {
-            this.buffer.push(b);
-        },
-        writeBytes: function (bb: number[]) {
-            Array.prototype.push.apply(this.buffer, bb);
-        },
-    };
+class TestOutputStream implements IOutputStream {
+    buffer: number[] = [];
+    writeByte(b: number) {
+        this.buffer.push(b);
+    }
+    writeBytes(bb: number[]) {
+        Array.prototype.push.apply(this.buffer, bb);
+    }
+}
+function createOutputStream(): TestOutputStream {
+    return new TestOutputStream();
 }
 
-t.testAsync("Write header of GIF89a", function (done) {
+t.testAsync("Write header of GIF89a", (done) => {
     var outputStream = createOutputStream();
     var gifWriter = new GifWriter(outputStream);
     gifWriter.writeHeader();
@@ -28,7 +26,7 @@ t.testAsync("Write header of GIF89a", function (done) {
     done();
 });
 
-t.testAsync("Write logical screen info without color table", function (done) {
+t.testAsync("Write logical screen info without color table", (done) => {
     var outputStream = createOutputStream();
     var gifWriter = new GifWriter(outputStream);
     gifWriter.writeLogicalScreenInfo({ width: 2, height: 1 });
@@ -41,7 +39,7 @@ t.testAsync("Write logical screen info without color table", function (done) {
     done();
 });
 
-t.testAsync("Write logical screen info with color table", function (done) {
+t.testAsync("Write logical screen info with color table", (done) => {
     var outputStream = createOutputStream();
     var gifWriter = new GifWriter(outputStream);
 
@@ -56,7 +54,7 @@ t.testAsync("Write logical screen info with color table", function (done) {
     done();
 });
 
-t.testAsync("Write table based image", function (done) {
+t.testAsync("Write table based image", (done) => {
     var outputStream = createOutputStream();
     var gifWriter = new GifWriter(outputStream);
     var indexedColorImage = new IndexedColorImage(
@@ -166,7 +164,7 @@ t.testAsync("Write table based image", function (done) {
     done();
 });
 
-t.testAsync("Write Loop Control (Application Extension)", function (done) {
+t.testAsync("Write Loop Control (Application Extension)", (done) => {
     var outputStream = createOutputStream();
     var gifWriter = new GifWriter(outputStream);
     gifWriter.writeLoopControlInfo(0x100);
@@ -184,7 +182,7 @@ t.testAsync("Write Loop Control (Application Extension)", function (done) {
     done();
 });
 
-t.testAsync("Write trailer", function (done) {
+t.testAsync("Write trailer", (done) => {
     var outputStream = createOutputStream();
     var gifWriter = new GifWriter(outputStream);
     gifWriter.writeTrailer();
@@ -194,5 +192,3 @@ t.testAsync("Write trailer", function (done) {
         "First 1 byte is byte of Trailer");
     done();
 });
-
-}).call(this);