+++ library/src/ts/MedianCutColorReducer.ts
@@ -1,52 +1,13 @@
-// same interface that `ImageData` interface of HTML standard
+import {selectKthElem} from "./Selection";
+
+// This interface corresponds to `ImageData` interface of HTML standard.
+// See: https://html.spec.whatwg.org/multipage/scripting.html#imagedata
 export interface IImageData {
     width: number;
     height: number;
-    data: number[];
+    data: Uint8ClampedArray | number[];
 }
 
-// Partition-based general selection algorithm
-// see : http://en.wikipedia.org/wiki/Selection_algorithm
-
-function swap(array: number[], idx1: number, idx2: number) {
-    var tmp = array[idx1];
-    array[idx1] = array[idx2];
-    array[idx2] = tmp;
-}
-function partition(a: number[], left: number, right: number, pivotIndex: number) {
-    var pivotValue = a[pivotIndex];
-    swap(a, pivotIndex, right);
-    var storeIndex = left;
-    for (var i = left; i < right; ++i) {
-        if (a[i] <= pivotValue) {
-            swap(a, storeIndex, i);
-            storeIndex = storeIndex + 1;
-        }
-    }
-    swap(a, right, storeIndex);
-    return storeIndex
-}
-function selectKthElem(list: number[], left: number, right: number, k: number) {
-    while (true) {
-        // select pivotIndex between left and right
-        var pivotIndex = Math.floor((right + left) / 2);
-        var pivotNewIndex = partition(list, left, right, pivotIndex);
-        var pivotDist = pivotNewIndex - left + 1;
-        if (k === pivotDist) {
-            return list[pivotNewIndex];
-        } else if (k < pivotDist) {
-            right = pivotNewIndex - 1;
-        } else {
-            k = k - pivotDist;
-            left = pivotNewIndex + 1;
-        }
-    }
-}
-
-function searchClosestColor(color: IColor, palette: IColor[]) {
-    var idx = searchClosestColorIndex(color, palette);
-    return palette[idx];
-}
 function searchClosestColorIndex(color: IColor, palette: IColor[]) {
     var min = 0;
     var found = false;
@@ -71,6 +32,7 @@ function searchClosestColorIndex(color: IColor, palette: IColor[]) {
     return (found ? foundIndex : closestIndex);
 }
 
+type ColorName = "red" | "blue" | "green";
 export interface IColor {
     red: number;
     blue: number;
@@ -110,14 +72,14 @@ class ColorCube {
         this.__maxB = maxB;
     }
 
-    divide() {
+    divide(): ColorCube[] {
         var cut = this.largestEdge()
         var med = this.median(cut)
         var r = this.divideBy(cut, med)
         return r;
     }
 
-    divideBy(cutTargetColor: string, median: number) {
+    divideBy(cutTargetColor: ColorName, median: number) {
         var list0: IColor[] = [];
         var list1: IColor[] = [];
         this.colors.forEach((c) => {
@@ -134,17 +96,17 @@ class ColorCube {
         }
     }
 
-    median(cutTargetColor: string) {
+    median(cutTargetColor: ColorName): number {
         var cc: number[] = [];
         var colors = this.colors;
         for (var i = 0, len = colors.length; i < len; ++i) {
             cc.push((<any>(colors[i]))[cutTargetColor]);
         }
-        var med2 = selectKthElem(cc, 0, cc.length - 1, Math.floor(cc.length / 2) + 1);
+        var med2 = selectKthElem(cc, Math.floor(cc.length / 2) + 1);
         return med2;
     }
 
-    largestEdge() {
+    largestEdge(): ColorName {
         var diffR = (this.__maxR - this.__minR) * 1.0;
         var diffG = (this.__maxG - this.__minG) * 0.8;
         var diffB = (this.__maxB - this.__minB) * 0.5;
@@ -186,11 +148,18 @@ class ColorCube {
     }
 }
 
+/**
+ * Simple color quantizer.
+ *
+ * It uses the median cut algorithm.
+ * If you have a full color image data and you want to write it as GIF by using GifWriter,
+ * you must do color quantization by using this class (or by other way) first.
+ */
 export class MedianCutColorReducer {
     private __imageData: IImageData;
     private __maxPaletteSize: number;
-    private __palette: IColor[];
-    private __colorReductionMap: { [colorRGBStr: string]: number; };
+    private __palette: IColor[] = [];
+    private __colorReductionMap: { [colorRGBStr: string]: number; } = {};
 
     constructor(imageData: IImageData, maxPaletteSize: number) {
         this.__imageData = imageData;
@@ -271,8 +240,11 @@ export class MedianCutColorReducer {
         return cubes;
     }
 
-    private __getLargestCube(cubes: ColorCube[]) {
-        var max: ColorCube = null;
+    /**
+     * @param cubes This must include one or more elements.
+     */
+    private __getLargestCube(cubes: ColorCube[]): ColorCube {
+        var max: ColorCube | undefined;
         var maxCount = 0
         cubes.forEach((x) => {
             var cc = x.getNumberOfColors();
@@ -281,6 +253,7 @@ export class MedianCutColorReducer {
                 maxCount = cc;
             }
         });
+        if (max === undefined) throw Error("`cubes` must have one or more elements.");
         return max;
     }
 }